+++ spr/src/commands/land.rs
@@ -68,7 +68,7 @@ pub async fn land(
 
     // Fetch current master from GitHub.
     run_command(
-        tokio::process::Command::new("git")
+        jj.git_command()
             .arg("fetch")
             .arg("--no-write-fetch-head")
             .arg("--")
@@ -253,7 +253,8 @@ pub async fn land(
 
     output("🛬", "Landed!")?;
 
-    let mut remove_old_branch_child_process = tokio::process::Command::new("git")
+    let mut remove_old_branch_child_process = jj
+        .git_command()
         .arg("push")
         .arg("--no-verify")
         .arg("--delete")
@@ -268,7 +269,7 @@ pub async fn land(
         None
     } else {
         Some(
-            tokio::process::Command::new("git")
+            jj.git_command()
                 .arg("push")
                 .arg("--no-verify")
                 .arg("--delete")
@@ -287,7 +288,8 @@ pub async fn land(
         // the merge might still not find the new commit.
         for i in 0..3 {
             // Fetch current master and the merge commit from GitHub.
-            let git_fetch = tokio::process::Command::new("git")
+            let git_fetch = jj
+                .git_command()
                 .arg("fetch")
                 .arg("--no-write-fetch-head")
                 .arg("--")